--- apollo-api/src/jvmMain/kotlin/com/apollographql/apollo/api/ScalarType.kt
@@ -1,10 +1,0 @@
-package com.apollographql.apollo.api
-
-/**
- * Represents a custom GraphQL scalar type
- */
-actual interface ScalarType {
-  actual fun typeName(): String
-
-  fun javaType(): Class<*>
-}
